requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-filter>=23.1.0
drf-yasg>=1.21.1
requests>=2.28.2
